find_unsafe2/cargo_subcommands: refactor: move most logic into lib.rs
